Improved naming convention a bit.
Due to changed naming, files are not recognized as 'modified'.

The naming format 'x_y_z-i_A_n_B_d_C' refers to:
x,y,z: The start, stepsize and end of the varying variable, which can be n_elements (1_3_300) or displacement (-06_006_06, meaning from -0.06, by 0.006 till 0.06).
A: The index within this set
B: The n_elements value
C: The displacement value (if not present, it's 0.02)